+++ main/index.js
@@ -1,6 +1,6 @@
 
 
-const { app, BrowserWindow } = require('electron');
+const { app, BrowserWindow, Menu } = require('electron');
 const { createMainWindow } = require('./modules/window');
 const { setupIPC } = require('./modules/ipc');
 const { setupAutoUpdater } = require('./modules/updater');
@@ -11,6 +11,24 @@ app.whenReady().then(() => {
   mainWindow = createMainWindow();
   setupIPC();
   setupAutoUpdater();
+
+  // Add View menu with theme toggle
+  const template = [
+    {
+      label: 'View',
+      submenu: [
+        {
+          label: 'Toggle Dark/Light Mode',
+          accelerator: 'CmdOrCtrl+Shift+D',
+          click: () => {
+            mainWindow.webContents.send('theme-toggle');
+          },
+        },
+      ],
+    },
+  ];
+  const menu = Menu.buildFromTemplate(template);
+  Menu.setApplicationMenu(menu);
 });
 
 app.on('window-all-closed', () => {